+++ src/main/java/com/diggydwarff/tobacconistmod/block/entity/TobaccoBarrelMode.java
@@ -1,0 +1,7 @@
+package com.diggydwarff.tobacconistmod.block.entity;
+
+public enum TobaccoBarrelMode {
+    IDLE,
+    FERMENTING,
+    AGING
+}