+++ src/main/java/com/diggydwarff/tobacconistmod/datagen/items/custom/LooseTobaccoItem.java
@@ -71,6 +71,22 @@ public class LooseTobaccoItem extends Item {
                     "Cut: " + TobaccoCuringHelper.getCutDisplayName(cutType)
             ).withStyle(ChatFormatting.GRAY));
         }
+
+        CompoundTag tag = stack.getTag();
+        if (tag != null) {
+            if (tag.getBoolean("Fermented")) {
+                tooltip.add(Component.literal("Fermented").withStyle(ChatFormatting.DARK_GREEN));
+            }
+
+            int aged = tag.getInt("AgedStages");
+            if (aged > 0) {
+                tooltip.add(Component.literal("Aged: " + aged).withStyle(ChatFormatting.GOLD));
+            }
+
+            if (tag.getBoolean("Ruined")) {
+                tooltip.add(Component.literal("Ruined").withStyle(ChatFormatting.RED));
+            }
+        }
     }
 
     @Override