+++ changelog.md
@@ -8,6 +8,9 @@
     - Air, sun, fire, or flue cure your tobacco!
     - Curing quality also tracked
     - Average-out leaf quality in the crafting grid to consolidate your production
+- NEW! Fermenation and aging system
+    - Added tobacco barrel block
+    - Manage temperature, humiditiy, and light levels to properly ferment and age your tobacco
 - NEW! Decoration and Quality of Life Features
     - Advancement system with custom tobacco-themed advancements
     - Multiple new paintings inspired by various tobacco cultural around the world